refactor: renamed module app into rule_engine
